% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{get_response}
\alias{get_response}
\title{Get API response}
\usage{
get_response(
  text,
  request_url,
  user_id = getOption("minhon.user_id"),
  token = get_token(),
  param = NULL
)
}
\arguments{
\item{text}{Text to translate}

\item{request_url}{API request URL}

\item{user_id}{User ID.}

\item{token}{Access token obtained by \code{\link[=get_token]{get_token()}}.}

\item{param}{Request parameters as named \code{list}}
}
\description{
Obtain API response from given API URL. Access
\url{https://mt-auto-minhon-mlt.ucri.jgn-x.jp/content/mt/} to get URL and
get details about the argument \code{param}.
}
\note{
Access token expires in 60 minutes.
}
