% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Get access token}
\usage{
get_token(
  api_key = getOption("minhon.api_key"),
  api_secret = getOption("minhon.api_secret")
)
}
\arguments{
\item{api_key}{API key}

\item{api_secret}{API secret}
}
\description{
Get access token from given API key and API secret.
Access \url{https://mt-auto-minhon-mlt.ucri.jgn-x.jp/content/mt/} to get
your API key and API secret.
}
\note{
Access token expires in 60 minutes.
}
