% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate a text with TexTra}
\usage{
translate(
  text,
  request_url,
  user_id = getOption("minhon.user_id"),
  token = get_token(),
  field = "/resultset/result/information/text-t"
)
}
\arguments{
\item{text}{Text to translate}

\item{request_url}{API request URL}

\item{user_id}{User ID.}

\item{token}{Access token obtained by \code{\link[=get_token]{get_token()}}.}

\item{field}{Field of response to obtain}
}
\description{
Translate a text with TexTra. Access
\url{https://mt-auto-minhon-mlt.ucri.jgn-x.jp/content/mt/} to get URL and
get details about the arguments \code{param} and \code{field}.
}
\note{
Access token expires in 60 minutes.
}
\examples{
\dontrun{
token <- get_token("API_KEY", "API_SECRET")
# English to Japanese
translate(
  "The quick brown fox jumps over the lazy dog",
  "https://mt-auto-minhon-mlt.ucri.jgn-x.jp/api/mt/generalNT_en_ja/",
  "USER_ID",
  token
  )

# You can store your API key, API secret and User ID as follows:
options(
  minhon.api_key = "API_KEY",
  minhon.api_secret = "API_SECRET",
  minhon.user_id = "USER_ID"
  )

translate(
    "The quick brown fox jumps over the lazy dog",
    "https://mt-auto-minhon-mlt.ucri.jgn-x.jp/api/mt/generalNT_en_ja/"
  )
}

}
